moved some plugins into the `plugins` package
