+++ backend/internal/compile/plugins/css.go
@@ -1,4 +1,4 @@
-package compilerServer
+package plugins
 
 import (
 	"bytes"
@@ -9,6 +9,7 @@ import (
 
 	es "github.com/evanw/esbuild/pkg/api"
 	"robinplatform.dev/internal/compile/buildError"
+	"robinplatform.dev/internal/httpcache"
 	"robinplatform.dev/internal/project"
 )
 
@@ -21,7 +22,7 @@ func wrapWithCssLoader(path string, css string) string {
 	}()`, path, css)
 }
 
-func getCssLoaderPlugins(appConfig project.RobinAppConfig) []es.Plugin {
+func LoadCSS(appConfig project.RobinAppConfig, httpClient httpcache.CacheClient) []es.Plugin {
 	return []es.Plugin{
 		{
 			Name: "load-css",
@@ -82,7 +83,7 @@ func getCssLoaderPlugins(appConfig project.RobinAppConfig) []es.Plugin {
 						return es.OnLoadResult{}, fmt.Errorf("failed to read sass file %s: %w", args.Path, err)
 					}
 
-					script, err := buildSass(args.Path, string(sass))
+					script, err := buildSass(httpClient, args.Path, string(sass))
 					if err != nil {
 						return es.OnLoadResult{}, fmt.Errorf("failed to build sass file %s: %w", args.Path, err)
 					}
@@ -97,7 +98,7 @@ func getCssLoaderPlugins(appConfig project.RobinAppConfig) []es.Plugin {
 	}
 }
 
-func buildSass(srcPath, sass string) (string, error) {
+func buildSass(httpClient httpcache.CacheClient, srcPath, sass string) (string, error) {
 	result := es.Build(es.BuildOptions{
 		Stdin: &es.StdinOptions{
 			Contents: fmt.Sprintf(`
@@ -117,9 +118,7 @@ func buildSass(srcPath, sass string) (string, error) {
 		Define: map[string]string{
 			"process.stdout.isTTY": "false",
 		},
-		Plugins: []es.Plugin{
-			esbuildPluginLoadHttp,
-		},
+		Plugins: LoadHttp(httpClient),
 	})
 	if len(result.Errors) > 0 {
 		return "", buildError.BuildError(result)
